useImageReader 훅 구현
